requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
